Made ComponentType be uint64_t

+++ platypus/ecs/components/Component.hpp
@@ -9,8 +9,7 @@ namespace platypus
 {
     class Scene;
 
-    // TODO: Make this enum class of uint32_t type!
-    enum ComponentType : uint32_t
+    enum ComponentType : uint64_t
     {
         COMPONENT_TYPE_EMPTY            = 0x0,
         COMPONENT_TYPE_TRANSFORM        = 0x1 << 1,
